requirements.txt
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.20
python-dotenv>=1.0.0
pydantic>=2.0.0,<3.0.0
langchain-core>=0.1.0
